remove code for embeddings relevancy calculation
I think it was only necessary because there used to be a token count
check. The token count check silently discarded files if the token count
was exceeded. To ensure that the most important files are tested against
the token count first, the removed code would calculate how similar the
open file is to the prompt. The more similar, the more important and
therefore would be "higher up" in the embeddings table.

We have removed the token count check now. It's all or nothing: either
token count wasn't hit and everything is good or the request fails
altogether. Nothing is silently discarded and there is no need to find
how relevant the file is to the prompt.