move state management out of copilot module
To be able to do modifications to both internal state and what's visible
in the UI, we need to keep track of the history state in one place. Chat
UI must be entirely derivable from the history state and copilot module
should be completely stateless - it should be just a wrapper for calling
copilot.