Rework system prompt logic
This change will track the system prompt that was used in the history.
To avoid hogging memory, the system prompt's name is stored. It acts as
a reference that can be used to figure out the entire system prompt
wherever needed.

This change will make it possible to keep using the same system prompt
after it has been changed and correctly revert to a previous one when a
prompt is reverted.

To make it clear to the user when a system prompt was changed, the
`/SYSTEM_PROMPT` commands are no longer hidden. Ideally in the future
there would be a notice of a change in the chat and/or they would be
highlighted.